added hard codded data for current customer until we have a way to track which customer is login, under MenuCustomerViewFragment added -addCartToDb purpose is to add the current cart to the database -clearcheckoutdatabase purporse is to clear the cart, the code is correct but the logic is wrong, after user press add to cart it clear the database need to implement this when the customer finalize the order
