+++ app/src/main/res/layout/fragment_cust_order_histroy.xml
@@ -1,14 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-<Button
-    android:id="@+id/button12"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:text="add to cart"/>
 
-</LinearLayout>
+
+</RelativeLayout>
